post vehicle fixed

+++ src/modules/user/user.service.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from "express";
 import { pool } from "../../config/db";
+import { JwtPayload } from "jsonwebtoken";
 
 const getAllUser = async () => {
     const result = await pool.query(`SELECT * FROM users`)
@@ -21,7 +22,12 @@ const updateUser = async (bodyData: any, userId: number) => {
 
     return result
 }
-const deleteUser = async (userId: number) => {
+const deleteUser = async (userId: number, user: JwtPayload) => {
+    const userBooking = await pool.query(`SELECT * FROM bookings WHERE customer_id = $1`, [userId])
+
+    if (userBooking?.rows.length) {
+        return { "message": "User not Deleted cause of active Booking" }
+    }
     const result = await pool.query(`
         DELETE FROM users
         WHERE id= $1`,
@@ -32,5 +38,5 @@ export const userServices = {
     getAllUser,
     updateUser,
     deleteUser,
-    
+
 }
